Made changes to improve controllers and service methods logic

+++ CocktailHeaven/CocktailHeaven/Controllers/UserCollectionController.cs
@@ -1,18 +1,19 @@
 ﻿using CocktailHeaven.Core.Contracts;
-using CocktailHeaven.Core.Models.Cocktail;
 using CocktailHeaven.Extensions;
 using Microsoft.AspNetCore.Mvc;
-using System.Security.Claims;
+using static CocktailHeaven.Infrastructure.Models.DataConstants.MessageConstant;
 
 namespace CocktailHeaven.Controllers
 {
 	public class UserCollectionController : BaseController
 	{
 		private readonly IUserCollectionService userCollectionService;
+		private readonly ICocktailService cocktailService;
 
-        public UserCollectionController(IUserCollectionService _userCollectionService)
+        public UserCollectionController(IUserCollectionService _userCollectionService, ICocktailService cocktailService)
         {
             this.userCollectionService = _userCollectionService;
+			this.cocktailService = cocktailService;
         }
 
 		public async Task<IActionResult> ShowFavourite()
@@ -45,25 +46,37 @@ namespace CocktailHeaven.Controllers
 
 		public async Task<IActionResult> AddToFavourite(int id)
 		{
+			if (await this.cocktailService.ExistsByIdAsync(id) == false)
+			{
+				this.TempData[ErrorMessage] = ErrorMessageCocktail;
+				return this.RedirectToAction("All", "Cocktail");
+			}
+
 			var userId = this.User.Id();
 
 			if(await this.userCollectionService.IsCocktailInFavouritesAsync(userId, id))
 			{
-				return RedirectToAction("ShowMore", "Cocktail", new { id });
+				return this.RedirectToAction("ShowMore", "Cocktail", new { id });
 			}
 
 			await this.userCollectionService.AddToFavouriteAsync(userId, id);
 
-			return RedirectToAction(nameof(ShowFavourite));
+			return this.RedirectToAction(nameof(ShowFavourite));
 		}
 
 		public async Task<IActionResult> AddToWishlist(int id)
 		{
+			if (await this.cocktailService.ExistsByIdAsync(id) == false)
+			{
+				this.TempData[ErrorMessage] = ErrorMessageCocktail;
+				return this.RedirectToAction("All", "Cocktail");
+			}
+
 			var userId = this.User.Id();
 
 			if (await this.userCollectionService.IsCocktailInWishListAsync(userId, id))
 			{
-				return RedirectToAction("ShowMore", "Cocktail", new { id });
+				return this.RedirectToAction("ShowMore", "Cocktail", new { id });
 			}
 
 			await this.userCollectionService.AddToWishListAsync(userId, id);
@@ -74,6 +87,12 @@ namespace CocktailHeaven.Controllers
 	
 		public async Task<IActionResult> AddToTried(int id)
 		{
+			if (await this.cocktailService.ExistsByIdAsync(id) == false)
+			{
+				this.TempData[ErrorMessage] = ErrorMessageCocktail;
+				return this.RedirectToAction("All", "Cocktail");
+			}
+
 			var userId = this.User.Id();
 
 			if (await this.userCollectionService.IsCocktailInTriedAsync(userId, id))
@@ -83,35 +102,52 @@ namespace CocktailHeaven.Controllers
 
 			await this.userCollectionService.AddToTriedAsync(userId, id);
 
-			return RedirectToAction(nameof(ShowTried));
+			return this.RedirectToAction(nameof(ShowTried));
 		}
 
 		public async Task<IActionResult> RemoveFromWishlist(int id)
 		{
 			var userId = User.Id();
 
+			if (await this.userCollectionService.IsCocktailInWishListAsync(userId, id) == false)
+			{
+				this.TempData[ErrorMessage] = ErrorMessageUserCollection;
+				return this.RedirectToAction(nameof(ShowWishlist));
+			}
+
 			await this.userCollectionService.RemoveFromWishListAsync(userId, id);
 
-			return RedirectToAction(nameof(ShowWishlist));
+			return this.RedirectToAction(nameof(ShowWishlist));
 		}
 
 		public async Task<IActionResult> RemoveFromTried(int id)
 		{
 			var userId = User.Id();
 
+			if (await this.userCollectionService.IsCocktailInTriedAsync(userId, id) == false)
+			{
+				this.TempData[ErrorMessage] = ErrorMessageUserCollection;
+				return this.RedirectToAction(nameof(ShowTried));
+			}
+
 			await this.userCollectionService.RemoveFromTriedAsync(userId, id);
 
-			return RedirectToAction(nameof(ShowTried));
+			return this.RedirectToAction(nameof(ShowTried));
 		}
 
 		public async Task<IActionResult> RemoveFromFavourite(int id)
 		{
 			var userId = User.Id();
 
+			if (await this.userCollectionService.IsCocktailInFavouritesAsync(userId, id) == false )
+			{
+				this.TempData[ErrorMessage] = ErrorMessageUserCollection;
+				return this.RedirectToAction(nameof(ShowFavourite));
+			}
+
 			await this.userCollectionService.RemoveFromFavouriteAsync(userId, id);
 
-			return RedirectToAction(nameof(ShowFavourite));
-
+			return this.RedirectToAction(nameof(ShowFavourite));
 		}
 	}
 }
